Move dequeue function and unit test into test resources folder
