+++ sts-mock/package.json
@@ -17,15 +17,20 @@
     "node": "20"
   },
   "devDependencies": {
+    "@aws-sdk/client-secrets-manager": "^3.734.0",
+    "@aws-sdk/credential-providers": "^3.734.0",
     "@eslint/js": "^9.19.0",
     "@types/aws-lambda": "8.10.147",
+    "@types/eslint__js": "8.42.3",
     "@types/jest": "29.5.14",
     "@types/js-yaml": "^4.0.9",
     "@types/node": "^20.17.14",
+    "aws-sdk-client-mock": "^4.1.0",
+    "aws-sdk-client-mock-jest": "^4.1.0",
+    "aws4-axios": "^3.3.14",
     "axios": "^1.7.9",
     "axios-retry": "^4.5.0",
     "aws-cdk-lib": "^2.177.0",
-    "aws-sdk-client-mock": "^4.1.0",
     "dotenv": "^16.4.7",
     "eslint": "^9.19.0",
     "jest": "29.7.0",
@@ -40,7 +45,10 @@
   },
   "dependencies": {
     "@aws-lambda-powertools/logger": "^2.13.0",
+    "@aws-sdk/client-dynamodb": "^3.734.0",
     "@aws-sdk/client-s3": "^3.735.0",
+    "@aws-sdk/util-dynamodb": "^3.734.0",
+    "@smithy/node-http-handler": "^4.0.1",
     "esbuild": "^0.24.2",
     "jose": "^5.9.6"
   },
