+++ sts-mock/src/functions/dequeue/getConfig.ts
@@ -1,4 +1,4 @@
-import { errorResult, Result, successResult } from "../utils/result";
+import { errorResult, Result, successResult } from "../../utils/result";
 
 export function getConfig(env: NodeJS.ProcessEnv): Result<Config> {
   if (!env.EVENTS_TABLE_NAME) {
