+++ sts-mock/src/utils/result.ts
@@ -7,8 +7,19 @@ type SuccessResult<S> = {
 
 type ErrorResult = {
   isError: true;
-  value: { errorMessage: string; errorCategory: ErrorCategory };
+  value: ErrorResultValue
+  // value: {
+  //   errorMessage: string;
+  //   errorCategory?: ErrorCategory;
+  // [key: string]: string | undefined;
+  // };
 };
+
+type ErrorResultValue = {
+  errorMessage: string;
+  errorCategory?: ErrorCategory;
+  [key: string]: string | undefined;
+}
 
 export const successResult = <S>(value: S): SuccessResult<S> => {
   return {
@@ -17,10 +28,7 @@ export const successResult = <S>(value: S): SuccessResult<S> => {
   };
 };
 
-export const errorResult = (value: {
-  errorMessage: string;
-  errorCategory: ErrorCategory;
-}): ErrorResult => {
+export const errorResult = (value: ErrorResultValue): ErrorResult => {
   return {
     isError: true,
     value,
